Update patch set 1
Patch Set 1: No score

This feels like such a hack.  I wonder if it wouldn't be better to try and do this inside of git fetch itself.  Where are we losing the connection?  Is it failing to actually open the TCP channel?  Or are we getting some data transfer than then aborting due to an evil man in the middle sending a RST packet?

Patch-set: 1
Reviewer: Shawn Pearce <1001240@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0