Update patch set 1
Patch Set 1:

@Shawn: the "setdefaultencoding" is for Python use.
It’s a necessary first step so that people can upload
with non-7bit characters. If you want to not limit to
one encoding, you would probably have to convert bet-
ween charsets in some place (the script?) anyway, as
a second step though.

That said, it’d probably be for the best if at least
the server-side encoding is standard Unicode (UTF-8).
Subversion does it that way too (conversion in the
svn client), no idea about git.

Patch-set: 1