Update patch set 7
Patch Set 7:

PTAL.  I now pass the list of projects that will be uploaded to the pre-upload hook.  Hopefully that is sufficient.

I have also added a fragment of XML and a fragment of python to the CL description to make it easy for someone to use this.

Patch-set: 7
Reviewer: Doug Anderson <1004907@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=0