Update patch set 3
Patch Set 3:

As indicated in earlier patches, I'm not an expert on how --depth works in git.  I'm relying on you and Anush for that knowledge and just acting as a code monkey in this CL.  ;)

I've adjusted things as you've indicated to only specify the --depth on the initial fetch.

Patch-set: 3
Reviewer: Doug Anderson <1004907@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=0