Update patch set 1
Patch Set 1:

I will be happy to make a split, but I still feel that I need some guidance with the --remove-on-error parameter since this is a flag that we both need and use in our automated environment as much as the --retry parameter. 

Would it be acceptable to only honor this parameter if there is an environment flag already set as well as renaming the parameter to "--remove-workspace-on-error"? Maybe remove the parameter altogether and only have it as an environment variable?

Patch-set: 1