Update patch set 1
Patch Set 1:

Yea I had the same thought reading this change, the platform stuff might be able to be a macro around the group stuff. There is a lot of similarity between the two.

I think the parts that matter are --platform auto and --platform all. auto meaning you have to determine the group name to use, while all meaning you have to include all candidate groups. But isn't that just not picking any group?

Patch-set: 1