Update patch set 2
Patch Set 2:

On a second thought, I agree that if we define the .gitmodules of the top of master branch as the blessed one among all versions in the history, this could save us much trouble.

But I still don't think parsing is the right choice when compared to registration. Here is my reason: We cannot specify project attributes and annotations if repo parses .gitmodules, and let submodules inheriting parent project's attributes cannot fix this problem. For example, it can be a real trouble when a sub-project needs sync-c but its parent project doesn't. You either set sync-c in parent project (and submodule inherits this) and then suffer from only sync'ing one branch/tag in parent project, or not set sync-c in parent project and suffer from very slow sync'ing sub-projects.

If you agree my reason, I will add this resolution to the commit log.

Patch-set: 2
Label: Code-Review=0