Update patch set 4
Patch Set 4:

It is, and in the case of http authentication, it's using the .netrc just like we will be when talking to the manifest server.

I think I understand what you are saying; something like "git is just an underlying tool, so let git take care of git, and we'll take care of repo".  Is that a fair assessment?  That's reasonable, but repo's relationship with git is fairly tight so I'm not sure that's best.

Also, the manifest server is such a small part of the sync process, it seems weird to have a dedicated --username,--password option for it.  What if there's a new kind of server we need to talk to in the future?  It's not clear what --username, --password is for.

Anyway, how do you feel about having --manifest_username, --manifest_password?

Patch-set: 4