Update patch set 4
Patch Set 4:

> something like "git is just an underlying tool, so let
> git take care of git, and we'll take care of repo".
> Is that a fair assessment? 

Yes, that's what I meant.  git is already capable enough of dealing with the necessary authentication, via .netrc or ssh keys, so there's no need for repo to interfere with it.

> Anyway, how do you feel about having --manifest_username,
> --manifest_password?

How about --manifest-server-username and --manifest-server-password?

And should they be added in this change? I was planning to add them separately.

Patch-set: 4