Update patch set 1
Patch Set 1: (2 inline comments)

"This is a good feature, but I'm not sure how I feel about all the prompting that will be going on."

Frankly I hate the prompting; I'm just not sure of a better approach, plus I didn't want to have to build an interactive "which remote would you like to configure" type menu system.

"If you have two remotes, you will now be prompted for 3 identities instead of 1. There are probably use cases with even more remotes. Not only that, but the first time it will be of the form name?/email? and the other times it will be of the form change?/name?/email?."

One potential; rather than requiring "go type your name in N times", track the credentials used- ask "if you'd like to enter a new identity in, press enter, else press \d+ of the previous identity to reuse here".  That's more interactive than I like, but still is scriptable.

"Another route might be that the manifest can dictate whether or not the user should be prompted for a separate identity. Chromium knows that it needs this, Android knows it does not, maybe they should get to say so in the manifest.
There are probably some other good ways to go about this too. Any ideas?"

Not really a fan of the "mark a remote to indicate if separate credentials should be used" angle- just seems hacky, pragmatic, but hacky.

Note we can bring the credential count for 2 remotes down from 3 to 2 also; the problem there is that then we lack a default to fall back to if a new remote shows up.  Potentially just tell people "reconfig/reinit" in that case?

There also is one other unstated flaw here; the user/email isn't updated if a project switches remotes.  Unlikely mind you, but it's there.  Same angle, the code doesn't know if it forced a value into a repo or if the user manually did- thus it can't update the repo.  Storing the user/email in repo.core.user, repo.core.name in each repo (and comparing that against the resolved value for that repo) is one way to detect if we're in control, or if the user has exerted control.

Patch-set: 1
Reviewer: Brian Harring <1011107@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=0