Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

I'm not sure why we are distinguishing between the time overview takes and the time info takes.  Would it not make sense to just print all the info that's locally available if some option isn't set and then print all the info that is locally and remotely available if said option is set?

Also, I think we should add a deprecation notice to the overview subcommand.

Patch-set: 3
Label: Code-Review=-1
Label: Verified=0