Update patch set 2
Patch Set 2:

> When exactly would this be used. We don't currently have a way to install repo outside of the source directory, do we?

For now, it's not possible to *use* Repo without depending on a Repo's git. From the Change-Id I7ccc5aa0630cf49ff24bcb67d4d7688663d3b890, it will be possible. Then, you're right, for now we don't have a way to install Repo standalone but it will be easy to do it (if needed, I will).

Anyway, this symlink patch fix some Repo use as said in the comment.

Patch-set: 2
Label: Code-Review=0