Update patch set 4
Patch Set 4:

Joe. You have a lot of good input and I'll try to address it all. Especially the code comments.

But, I must address your first comment here.
"Anyone who is working on an even remotely fast-paced project, and might have a lot of local commits spread across different projects, is going to find this command very hard to read and hard to work with."

This command is not an overview command, this command is not for getting an overview of all the commits across all the different projects. It was never meant to be that. It was meant to give information about the complete repository. Which is does. 

Your overview command is excellent for what it does. I don't see how you can get the impression that the info command is trying to do the same thing. The stuff printed by a bare "repo info" command is extremely useful to me and other here at Sony who work with a lot of different branches on the manifest level but in the same base folder. Lets just say that there is often at least +4 manifest branches for the same project and we need to move around them quickly and also sometimes get a reminder of which one we are on. This is for the 2 first information posts.

The other ones are there because it's also good to get an overview of the current git-repo you are working in. Are we on a sha1, where would I push, will this affect more than one project etc. All things that usually come from inspecting the manifest file. But now easily available in one command.

Now. That is for the basic command without any options. The -a flag appeared when I saw the need to get a diff between the current status of my git-repo and the upstream. It was _never_ meant to give an overview of all commits in all projects. It was meant to give the status in the current git. Where it can be useful to see how far ahead the upstream is. 

So, lets start with the numbered comments you have:

1) I am not sure what the argument here is. Do you want me to rename the flag to something more "networky". The functionality displayed by -a requires network access. This makes is a lot slower yes. That is why it's optional.

2) Is something not working with the coloring? Have I done it wrong? You obviously see that I try to color this since you've commented on the coloring code. Please let me know if and why (if you know) it doesn't work.

3) Again... this is not the point of the command. Please try to understand. It was the first thing I wrote as a comment. It's all about scope. This command does not work on the scope you think it does. It is not the overview command, it doesn't try to be, it never will be.

After the tl;dr:
This can of course be discussed. I don't see the harm in doing something easier for a user, remember. It's optional. repo sync --fetch-only will fetch for all projects which will take a lot longer time. My current working repo has 522 projects in it. The reason to hit the network is to get the diff. To know how far behind the upstream you are. If you do not find that useful, do not use it. This would be far from the only command in repo to gather more than one command into another.

Coloring:
Not sure what the problem is here. I must have misunderstood how the coloring works. Please let me know how to do it.

Patch-set: 4
Label: Code-Review=0