Update patch set 4
Patch Set 4: I would prefer that you didn't submit this

Marking as -1 as per my other comment, quoted below:

"""
I'm not sure I understand what the problem is. Using a clean repo (without this patchset) I removed my .netrc entry and tried to upload a change and was prompted for my username and password. In what situation does this patchset offer a prompt where it wouldn't be offered already?
"""

Patch-set: 4