Update patch set 4
Patch Set 4: (6 inline comments)

Ok, answered and fixed some stuff regarding your comments, on to the other stuff:

(Note, -a is changed to -d in latest patch)

> My point is that the -a functionality actually *doesn't* "require" network access: it only requires network access to get an up-to-date version. And for that, I am saying that it would be better to let the user decide when and if that network access actually occurs. Thus the "repo sync" suggestion.

> I understand that, but I still think it's a bad thing to force the network access on the user, without letting him decide. The user will *know* that the information will be out of date until the next repo sync. I don't think that's a bad thing, and it is in line with the functionality of both repo and even git itself

Here I disagree, first, it's not 'forced', it's triggered by the -d flag. The information regarding this is in the help for the command, it explicitly states that you'll hit the network. Second, the usefulness of this command is that it does hit the network, and only for the current project.

Third, arguing that you can get this info with other commands is a moot point. All of repo can be done with other commands, the whole purpose of repo is to simplify not to make aliases for git commands.

Anyway, take a look at the new code and try it out. Check the coloring and see what you think about the arguments.
Should be patch set 7 ;)

Cheers

Patch-set: 4