Update patch set 7
Patch Set 7:

> Here I disagree, first, it's not 'forced', it's triggered by the -d flag. The information regarding this is in the help for the command, it explicitly states that you'll hit the network. Second, the usefulness of this command is that it does hit the network, and only for the current project.

I think I'm not explaining myself properly.  I know that the remote commits are only printed if you give the --diff option.  But I'm saying that you can still show a local representation of that information (`git log --oneline HEAD..gerrit/master`), without hitting the network.  It may be out of date, but that's also true of everything else related to git.  I'm saying the --diff output is useful information to have, but you can still print that information using local refs only, without requiring the network sync.  That is, it "can" be done locally, but the `info --diff` command *forces* it to hit the network.  That's my only point.  I like the information that the --diff output shows, but I wouldn't want it to hit the network every time.

Especially since I'm most likely to look at that report at the start of the day, and/or the end of the day.  And it's at the same times that I am also doing a `repo sync` and uploading my commits for review.  So the local refs are already generally up to date by that point.  That's an isolated case, I know -- and I'm not saying that my situation matches everyone's.  But it certainly matches *someone's*.

> Third, arguing that you can get this info with other commands is a moot point. All of repo can be done with other commands, the whole purpose of repo is to simplify not to make aliases for git commands.

You're right.  Given that, if the command had a --local-only/-l option that could bypass the network sync, it would be convenient.

Patch-set: 7
Label: Code-Review=0