Update patch set 1
Patch Set 1:

I guess I'm just not sure when this gets called.  In the other change, the code was obviously called at the end of every sync, for this one I need to hunt around for square brackets after project variables which is a little harder to do.

Given your comment, I'm guessing that this one happens during every sync as well.

Patch-set: 1