Update patch set 6
Patch Set 6: Looks good to me, but someone else must approve; Verified

It works as expected.  If files are added in local_manifests, they are included in the sync.

If local_manifest.xml exists, it is added.

It also works if local_manifest.xml exists at the same time as manifests in local_manifests, assuming gits are not duplicated.

Patch-set: 6
Label: Code-Review=+1
Label: Verified=+1