Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

This will need to be rebased; couple of changes have gone in that add more prints.

Would be good to get this in soon, there are likely to be more changes going in that add prints.  If this goes in first we can start to use the new style.

Patch-set: 2
Label: Code-Review=-1