Update patch set 2
Patch Set 2:

Sorry for confusion. If a submodule is registered, it is better registered at paths that match paths in .gitmodules so that repo and git are on the same page.

And if a submodule is not registered, repo tries to parse .gitmodules and thus should be always on the same page with git.

(The benefits of registering, given the cost of maintaining matched paths, is that you may override manifest attributes.)

Patch-set: 2
Label: Code-Review=0