Update patch set 5
Patch Set 5:

David Pursehouse, i had forwarded an email to you with two attached files to make clear why i change it from 'name' to 'path', i means, we can place the git project in a special path. Indeed, if the git path attr not exist in the manifest file, its value will be the same as the git name attr.

Patch-set: 5