Update patch set 5
Patch Set 5:

This will break repo for people who use the --mirror option and expect the directories to be named according to the name of the git (even if the path attribute is specified).

If this functionality is needed, I would prefer it to be enabled by a command line option.  The default behaviour should be as it is now, and only changed if the user explicitly passes the command line option.

Patch-set: 5