Update patch set 9
Patch Set 9:

To be clear on functionality, to use this you need to use --force-path when creating the mirror, then again when creating the working tree that uses the mirror, correct?  If so, that seems slightly less than ideal.

Can you give me a specific example of a path name conflict this solves?

Additionally, I remember considering this problem prior to the hackathon .  I was thinking that it'd be good to store each manifests set of projects in a different project subdir.  So something like/projects/local_manifest1/<project name>.git.  I'm not sure how we could go about this without breaking existing tools that rely on the current structure.  Since not many people are using multiple local manifests yet, perhaps one way would be to create a local_manifest_projects directory beside the projects directory and store those there.  Anyway, then there would only be one way for this to happen and we wouldn't need a flag to toggle.  However, I'm not sure it's worth the effort.

Thoughts?

Patch-set: 9
Reviewer: Conley Owens <1006257@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0