Update patch set 1
Patch Set 1:

(3 comments)

Replies inline. I'll upload a new version of the patch shortly.

We're considering using Repo to manage multiple repositories (obviously). It's very important that when a new user does "repo start" (after a "repo sync") they get a good state *across all repositories*. At the same time, we can't ensure that a code review on two different repositories will complete at the same time.

So, what we're trying to guard against is:
- Changeset is created for repos "firstrepo" and "secondrepo"
- Change is approved for "firstrepo"; remains in code review for "secondrepo"
- User does "repo sync; repo start" and geta change based off the newer "firstrepo" and the older "secondrepo". In other words, an inconsistent state.

The only way I could see to do this is to have all changes based off one branch and submitted for review on another. Then, when the changes pass CR and are *all* merged into develop, you push the consistent-across-all-repositories state atomically to master.

If we can't set this property in the manifest file it's substantially less useful, as that would mean users would have to remember to do "repo upload -D develop" every... single... time they upload a change. But I did add the option of setting an upload destination on the command line as part of this patch; that functionality by itself is better than nothing.

Patch-set: 1
Label: Verified=0