Update patch set 6
Patch Set 6:

Sorry to reply so late...

> what exactly is the use case for this?

The use case is as mentioned in the commit message.  A user having several workspaces, but wanting to use the same local manifests in each one, has to copy the files into each workspace.

This isn't a critical feature; it's just something I thought would be nice to have.  I can abandon it if nobody else thinks it's needed.

Patch-set: 6