Update patch set 6
Patch Set 6:

Yea, I don't think this is good if you have to remember the flag each time.  It's a very different workflow then how repo typically works (being in a static state until the next time you init).

You'd need to add this flag to forall and any other command that looks at the manifest (which is a lot of them).

Patch-set: 6