Update patch set 9
Patch Set 9:

This is very, very nice. But I am not sure why the git template is limited to just hooks? Would it not be a more natural fit if all of the git repositories in a manifest shared the same git template, and that includes the files in .git/hooks as well as the files in .git/info such as the files "exclude" [see gitignore(5)] and "attributes" [see gitattributes(5)]?

Maybe for the encore, but in that case "projecthook" is perhaps not the best tagname to start with. I would shamelessly suggest replacing it with "git-template" to remove any confusion over what this tag is used for. Otherwise there is likely to be some confusion over "projecthook" visavi "repo-hooks". The latter I might add is not very well known as it is quite new, and to top it off there seems to be zero documentation for how it is used.

Patch-set: 9
Reviewer: Mats Bengtsson <1012267@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0