Update patch set 5
Patch Set 5:

> If that's the case, why don't you just set clone-depth in the
 > manifest or use the --depth flag?

Correct my if I'm wrong, but it is my understanding that the depth option is from the top of a branch. We need this behavior for any revision in a project.

Is there another way to get any revision without its surrounding history in git? Would this other way be applicable to repo?

Patch-set: 5
Reviewer: Julien Campergue <1013723@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0