Update patch set 1
Patch Set 1:

In order to communicate the progress and fetch times back to the frontend process, we'll need to set up some queues. This will basically involve rewriting progress.py and the _FetchTimes class.

I think that'd be a win, but it'd be a big patch -- too big, I think, for one patch.

This patch simplifies the code to handle jobs == 1 and jobs > 1 the same way, which helps us get towards our end goal, so I'd suggest we consider pushing this one as-is.

Thoughts?

Patch-set: 1