Update patch set 4
Patch Set 4:

> > I can modify the patch to avoid fetching tags in the initial case
 > > if you like?
 > 
 > Yes, I think that would be better.   Thanks.

I looked into this a bit, and I don't think this is safe to do.

On the initial fetch, if you don't fetch tags, you can easily end up not fetching the commits referred to in the manifest if they're not part of any branch's history. For example, this manifest fails to initialize properly when no tags are fetched:

<?xml version="1.0" ?><manifest>
  <remote fetch="https://git.mozilla.org/external/caf" name="caf"/>
  <project remote="caf" name="device/sample" path="device/sample" revision="237bd668d0f114d801a8d6455ef5e02cc3577587"/>
</manifest>

This is because 237bd668d0f114d801a8d6455ef5e02cc3577587 isn't contained in any of the branches, it's only referred to by a tag.

The original purpose of this patch is to prevent fetching all the refs for subsequent sync operations if we already have the specified revision locally.

Patch-set: 4
Label: Verified=0