Update patch set 1
Patch Set 1:

(1 comment)

According to my `man curl` this option was added in 7.20.2. Slightly confusing that http://curl.haxx.se/docs/releases.html does not list this release but based on neighboring releases I would guess somewhere in the April-June 2010 range. I don't think we need to support curl older than that, but just thought I would mention it.

Also I believe you mentioned an error message suggesting it is possible to compile curl without support for file://; is this option safe to use with such a version? If not, ideally we would not want to pass it unconditionally. It may also be that some protocols are optional but file:// is not (or essentially always compiled in), or that it's too much work to try to detect it.

Patch-set: 1