Update patch set 5
Patch Set 5: Code-Review-1

Sorry for the delay.  I think this adds a layer of complexity to the configuration that isn't needed.  Manifests aren't really supposed to be entirely independent/agnostic of their greater context and I don't think the burden involved in creating a manifest containing the necessary prefixes is that burdensome.

Patch-set: 5