Update patch set 1
Patch Set 1:

Shawn, what is the best way to --unshallow a git repository if you don't know what refspec was specified in the shallow fetch?  This is a super ugly hack that stores the last shallow fetch command in a projects git config, but there's got to be a better way.

Patch-set: 1