Update patch set 5
Patch Set 5:

> (1 comment)
 > 
 > Hello, thanks for replying.
 > 
 > repo doc (manifest-format.txt) states that included manifests "must
 > be a usable manifest on its own" and we have base our projects
 > structure on it. For completeness, I asked coworker that use it to
 > explain how and why dir_name is currently used.
 > 
 > Thanks.

Hello all,

Actually we work on a "big" project that acts as a compilation of standalone projects.

Each project has their build/configuration method implemented at the root of it (mostly through copyfile). The main advantage of having the dir_name implementation is to allow us to recover any project and to build/configure them without having intermediate checks as we know that whatever the project is, the needed tool is at the root.

It just needs a little customization for integrators to make sure that tools options are uniformed but for the developer it is quite a nice feature to have the simplest compilation method available.

Thanks for reading,

Best regards.

Patch-set: 5
Reviewer: Julien Thomas <1016127@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0