Update patch set 1
Patch Set 1:

> Fragile, yes, but repo itself is a brittle tool at the moment and can't really afford to have this kind of functionality added for such a small use case.

Having a downstream copy of a repo-managed project is not a small use case; in particular, together with the subsequent change to enable multiple sets of repositories based on groups, this makes it feasible to work on a repo-managed project without having to create an internal branch of its upstream manifest and keep that branch up to date.  "fragile" is an understatement here; this change would eliminate a massive amount of engineering and logistical effort for developers working on Android and Chrome OS.

This is a very small change, consisting of half documentation, half a single addition of an if-case to parse this new element, and one minor refactor to take a copy/paste duplicated line to parse groups and make it a function.  (If you'd like, I can split that last one out into its own commit, which seems worthwhile regardless.)

Is any particular part of this change concerning you?  I'd be happy to redesign it further if you believe it introduces undesired complexity.

I have a test case I could submit as well, showing exactly how to use this feature, with a manifest and some sample commands; would that help?

Patch-set: 1