Update patch set 1
Patch Set 1:

I appreciate the efforts, but I just don't think this would be a healthy change for repo.

> maintain multiple independent branches of a manifest, one per target set of repositories; those branches would share many common elements

Yes, that's how it supposed to be done.  Right now there are about 4 too many ways to slice and dice repo projects.  It sounds like what you are looking to do is remove the need for manifest branches and replace it with a pure group solution.  That's an OK idea, but the issue is that in practice it will just make things more confusing and make there be far more options than there should be.

If we were to actually move in this direction, there would have be a consensus that this is really where we wanted to go.  I don't think it's wise to just throw this out there as another option.

Furthermore, my hope is to eventually migrate Android to a more pure git submodules solution, in which case the groups feature is going to have to be marginalized.

Patch-set: 1