Update patch set 5
Patch Set 5:

Hello,

 > Does the manifest subcommand continue to work appropriately? 
We did not encounter problems so far with our use cases. Note that it still generates a "flat" (unique) manifest (without any include, pathes are then prefixed with 'dir_name').

 > How much testing have you done?
We use it daily for our own projects in our corporate repo. We probably do not cover every use case though.

About the patch set 6, it fixes the parse order which was not kept in previous patch sets.
It has also been rebased on master.

Thanks

Patch-set: 5