Update patch set 3
Patch Set 3: Code-Review-1

This is quite a fundamental change in the behaviour of subcmds/sync.py that I'd rather see it being an option, rather than the default behaviour - arguably as, I imagine, most users use it to simply fetch source code, and are not really interested in anything inside the git repositories themselves. Either way, there should be an option to turn this on or off.

Patch-set: 3