Update patch set 3
Patch Set 3:

In our case the git that is added in the local manifest has some new commits on the remote, which need to be fetched.

I'll follow up wit this next week to get the exact steps to reproduce it.

In the meantime I don't think it should block this commit.

Patch-set: 3