Update patch set 3
Patch Set 3:

Here's how you can reproduce the problem.

First, try this:

$ git clone https://android.googlesource.com/platform/framworks/base

Observe that there is a progress monitor. This repo is >1G, so depending on your network connection it may take a couple of minutes to hours to finish. Feel free to Ctrl-C it.

Now, let's try cloning that repository using repo.

 REPO_PATH=<path to your local checkout of repo>
 REPO_BRANCH=<name of your development branch containing this change>
 
 git init test-manifest
 cd test-manifest
 cat > default.xml << EOF
 <?xml version="1.0" encoding="UTF-8"?>
 <manifest>
   <remote name="test" fetch="https://android.googlesource.com/" />
   <default revision="master" remote="test" />
   <project path="frameworks/base" name="platform/frameworks/base" />
 </manifest>
 EOF
 git add default.xml
 git commit -m "test manifest"
 cd ..
 mkdir test-repo
 cd test-repo
 repo init --repo-url=$REPO_PATH --repo-branch=$REPO_BRANCH --no-repo-verify -u ../test-manifest
 repo --trace sync -j 1 --no-clone-bundle

The last line of output I see is:

 : git fetch test --tags +refs/heads/*:refs/remotes/test/* +refs/heads/master:refs/remotes/test/master 2>|

Then it hangs. If you wait long enough, it'll finish, but it's not very useful in the meantime.

Considering that a typical sync of Android can take minutes to hours, we really need to show that progress monitor.

Patch-set: 3