Update patch set 7
Patch Set 7: -Code-Review

Is there any concern with always capturing the output?  Since certain git commands will change how they output, since git changes the output depending on if it thinks the output is a TTY or not.  Can probably use the '--progress' flag to fix that, but would need to search for all cases.  Or modify GitCommand to add '--progress' to certain git sub-commands.  Thoughts?

Patch-set: 7
Label: Code-Review=0