Update patch set 8
Patch Set 8:

> Personally, I would use a Queue for all output. The code would be
 > much simpler, be compatible with Windows, and, at a later stage,
 > could be used for anything that outputs (can't remember what does
 > off the top of my head).

I might be missing something but the goal is to do nonblocking reads from the underlying stderr file descriptor opened by the subprocess object, right? Once you pump the data into a Queue a consumer could read it in a nonblocking manner, but how are you proposing to get the data into the Queue in a manner that is both nonblocking and platform-independent?

Patch-set: 8
Label: Verified=0