Update patch set 3
Patch Set 3:

I guess this is not a normal situation.  I think in most use cases the manifest that is being mirrored is also being used as a regular manifest, and thus any problems related to revisions would be exposed (as expected) when syncing it normally.

In our case we're using (or abusing?) the smart-tag feature to dynamically generate a mirror manifest. The manifest itself is empty (i.e. no projects are listed) and the projects are added by the manifest server based on the value given with the -t option.  Such a manifest is only used for syncing the mirror, never for a workspace.

Patch-set: 3