Update patch set 1
Patch Set 1:

I can't think of anything it will break. I was more concerned about the change in behaviour. 

If you think it's ok, can you amend it to remove the RFC and merge it for the release you're doing today? I'm not at the office today.  Thanks.

I've run it and it works as I expected.

Patch-set: 1