Update patch set 3
Patch Set 3:

(1 comment)

> I think the problem is the "git fetch" in the README.md file:
 > 
 > https://github.com/winksaville/sel4-helloworld-test-linkfiles-manifest/blob/master/README.md
 > 
 > was pointing at patch set 2:
 > 
 > git fetch https://gerrit.googlesource.com/git-repo
 > refs/changes/41/68241/2 ...
 > 
 > when it should have been pointing at patch set 3:
 > 
 > git fetch https://gerrit.googlesource.com/git-repo
 > refs/changes/41/68241/3 ...
 > 

No, I already took that into account and manually fetched patch set 3.  I had to manually do that line anyway because fetches from gerrit.googlesource.com are not working for me (see issue 2431).


 > I've fixed that. I've also linked in a python script to actually
 > test the result.
 > 
 > So please try it again.

I've figured out what the failing use case is.  If I run through the tests with a completely new workspace folder, then it works.

But in my tests before, rather than completely removing the helloworld-test-linkfiles-2 folder and making it again, from inside that folder I had done `rm -rf *` and `rm -rf .repo`.

I realised today that these commands do not result in removal of the `.build` folder.  When that folder is there, the subsequent repo sync does not create the symlinks, and the result is:

 $ ../../test_LinkFile.py ../..
 Errors in directory ../..
   "Kbuild":".build/Kbuild", was missing from foundEntries
   "Makefile":".build/Makefile", was missing from foundEntries
   "Kconfig":".build/Kconfig", was missing from foundEntries
   "newapp.sh":".build/newapp.sh", was missing from foundEntries
 1 error

After completely removing the workspace and starting from scratch, it works as expected:

 $ ../../test_LinkFile.py ../../
 Success

So now the question is, is it supposed to work even when the .build folder from a previous sync was not removed properly?

Patch-set: 3