Update patch set 2
Patch Set 2: Code-Review+2

> Hrmmm... David, should I click this "Rebase" button? I'm new to
 > Gerrit and I'm not sure if this button only shows up if it is
 > necessary.
 > 
 > http://gsc.io/s/rebase_1B6AA3EE.png

The button is always there.  When you click it you get a modal dialog which allows you to rebase either onto the tip of the branch or to another parent.

In case the commit is already based on the tip of the branch, the "Rebase" button in the dialog is disabled until you use the option to change the parent.

Patch-set: 2
Label: Code-Review=+2