Update patch set 3
Patch Set 3:

It just got released a few hours ago, and it's just above the exception backtrace. Not very clear, but it was better than nothing.

That rewording would be a good start, but at least some of the people hitting this problem are going to have it across many projects. Any project originally synced with <v1.12.8 will need to be force synced.

Patch-set: 3
Label: Verified=0