Update patch set 1
Patch Set 1:

> Have you been able to test it?

I did run a basic test of this with a manifest file that contained dest-branch attributes on a default element as well as on several projects, and the attribute seemed to be output properly in that case. Specifically, the output didn't include dest-branch attributes on projects that were inheriting a value from the default element.

Patch-set: 1