Update patch set 4
Patch Set 4:

I was thinking the combination of the error message and a flag named --force should be enough warning to look at the help documentation before proceeding, but it certainly wouldn't hurt to expand the message. I've uploaded a revised patch that's a bit more explicit.

Patch-set: 4
Label: Verified=0