Update patch set 3
Patch Set 3:

@Dan Willemsen, Thank you for telling me that revert commit.
In most teams, manifest owner is trusted and permissions controlled, so I set user_allows_all_hooks(the first parameter of hook.Run func) to True, which will never prompt about running the hook.And it seems to be not very well if we prompt user when they excute repo sync cmd.
But as the description of that revert commit, if we assume manifest owner is a offender, my commit have the same problem.
How to balance security and practicability? Any suggestion?

Patch-set: 3